Internal deploy python client functionality (#218)
The client operates in two modes, external mode (default) and self-hosted mode.

Also add a few extra functions for the missing API calls (e.g. delete bundle/endpoint, put endpoint)

For uploading model bundles in self-hosted mode, there is an upload_bundle_fn that needs to be passed to the client. This allows clients to specify their own model bundle storage backends. An example implementation is

def upload_bundle_fn(serialized_bundle, bundle_url):
    transport_params = {"session": Session(profile_name="ml-worker")}
    with smart_open.open(bundle_url, "wb", transport_params=transport_params) as f:
        f.write(serialized_bundle)
n.b. the code assumes that the REST APIs for the public route and the internal route are nearly identical, e.g. creating a model bundle is always at POST /model_bundle/ for example, and it takes in a JSON with a key url.